[Fix] migrate create_manuals.rb

+++ db/schema.rb
@@ -39,8 +39,8 @@ ActiveRecord::Schema.define(version: 2021_09_09_164328) do
   end
 
   create_table "bookmarks", force: :cascade do |t|
-    t.bigint "user_id", null: false
-    t.bigint "manual_id", null: false
+    t.integer "user_id", null: false
+    t.integer "manual_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["manual_id"], name: "index_bookmarks_on_manual_id"
